requirements.txt
django==1.4
#file:///Users/michaelrooney/Downloads/Django-1.4.tar.gz
python-dateutil
